requirements.txt
pymumble
python-decouple
numpy
Pillow
